+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.0
+
+* feat: support record items with custom attributes.
+
+### Pipeline version compatible
+Need use pipeline version 1.1.x and above
+
 ## 0.1.0
 
 * feat: add basic android and swift SDK APIs.